# Reduce GC frequency to run tests faster