--- a/test/ballast_test.go
+++ b/test/ballast_test.go
@@ -3,7 +3,7 @@ package test
 var ballast []byte
 
 func init() {
-	// 100MiB memory ballast to reduce the frequency of GC
-	// reduces runtime of tests by 20%
-	ballast = make([]byte, 100<<20)
+	// Allocate 200MiB memory ballast to reduce the frequency of GC, which reduces
+	// runtime of tests by ~20%.
+	ballast = make([]byte, 200<<20)
 }